==> notebooks/1_modality_specific/symptomatic/01_kmeans_clustering.ipynb <==
import config
import os
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import display
import datetime
from utils.data_exploration_utils import investigate_data

today = datetime.date.today()

base_dir = config.RAW_DATA_PATH
proc_dir = config.PROC_DATA_PATH
save_dir = os.path.join(proc_dir, f"{today}_kmeans_clustering")
os.makedirs(save_dir, exist_ok=True)

df_name = "inmodi_data_personalinformation_unpivoted.csv"
folder_name = "2025-07-03_data_exploration"

df = pd.read_csv(os.path.join(proc_dir, folder_name,df_name))

# Preprocessing

## One-Hot Encoding

categorical_column = 'gender'

col = ['pain', 'age', 'ce_height', 'ce_weight',
       'ce_bmi', 'ce_fm', 'gender', 'OKS_score', 'UCLA_score',
       'FJS_score', 'KOOS_pain', 'KOOS_symptoms', 'KOOS_sport', 'KOOS_adl',
       'KOOS_qol']

df_encoded = pd.get_dummies(df[col], columns=[categorical_column], drop_first=True)
df_encoded['gender_male'] = df_encoded['gender_male'].astype(int)

investigate_data(df)

- remove nan values where all scores are missing
- do one round without Koos and one with, bc we have some rows were all of these are missing.
- maybe one round just with bmi fm etc.
